+++ next.config.ts
@@ -1,8 +1,10 @@
 import type { NextConfig } from "next";
 
- 
+
 import { initOpenNextCloudflareForDev } from "@opennextjs/cloudflare";
 initOpenNextCloudflareForDev();
+
+// https://nextjs.org/docs/app/api-reference/config/next-config-js/ppr
 
 const nextConfig: NextConfig = {
   /* config options here */
@@ -11,6 +13,9 @@ const nextConfig: NextConfig = {
     fetches: {
       fullUrl: true,
     },
+  },
+  experimental: {
+    ppr: true
   },
   images: {
     remotePatterns: [
