+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
+    "dev-throttle": "THROTTLE_REQUEST=true next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -13,7 +14,6 @@
     "deploy": "npm run build-deploy && opennextjs-cloudflare deploy",
     "cf-typegen": "wrangler types --env-interface CloudflareEnv cloudflare-env.d.ts",
     "create-kv": "npx wrangler kv namespace create"
-
   },
   "dependencies": {
     "@opennextjs/cloudflare": "^1.5.1",
@@ -34,7 +34,7 @@
     "js-cookie": "^3.0.5",
     "lucide-react": "^0.501.0",
     "motion": "^12.18.1",
-    "next": "^15.3.4",
+    "next": "^15.4.2-canary.9",
     "next-sanity": "^9.12.3",
     "postcss": "^8.5.3",
     "react": "^19.1.0",
@@ -53,7 +53,7 @@
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^9",
-    "eslint-config-next": "^15.3.4",
+    "eslint-config-next": "^15.4.2",
     "tailwindcss": "^4.1.4",
     "tailwindcss-animate": "^1.0.7",
     "typescript": "^5",
