+++ src/app/(home)/portfolio/[project_type]/[slug]/page.tsx
@@ -1,37 +1,23 @@
-import { ProjectView } from '@/components/layout/project/projectView';
-import { ProjectProvider } from '@/redux/provider/projectProvider';
-import { Routes } from '@/types/routes';
-import { redirect } from 'next/navigation';
-import React from 'react';
-import { getCookie } from '@/actions/cookies/cookieHelper';
-import { isProjectLiked } from '@/lib/utils';
-import { ProjectsModel } from '@/models/projectsModel';
+import React, { Suspense } from 'react';
+import { ProjectMainPageProvider } from '@/components/layout/project/projectMainPage';
+import { Skeleton } from '@/components/ui/skeleton';
+import { LoaderCircle } from 'lucide-react';
 
 type ProjectPageProps = {
   params: Promise<{ slug: string; project_type: string }>;
 };
 
-const HOME_ROUTE: Routes = '/portfolio';
-
 export default async function ProjectPage({ params }: ProjectPageProps) {
   const { slug } = await params;
 
-  const [project, likedProjects = []] = await Promise.all([
-    ProjectsModel.getInstance().getProject(slug),
-    getCookie<string[]>('likes')
-  ]);
-
-  if (!project) {
-    redirect(HOME_ROUTE);
-  }
-
-  const isLiked = await isProjectLiked(project.slug, likedProjects);
-
   return (
-    <ProjectProvider project={project} isProjectLiked={isLiked}>
-      {/* div used as the background, used here to allow the two div elements to slide */}
-      <div className="relative flex-1 bg-black overflow-y-scroll" />
-      <ProjectView />
-    </ProjectProvider>
+    <Suspense fallback={
+      <Skeleton className='w-full h-full bg-zinc-900 flex items-center justify-center'>
+        <LoaderCircle className="animate-spin" />
+      </Skeleton>
+    }>
+
+      <ProjectMainPageProvider slug={slug} />
+    </Suspense>
   );
 }
