+++ src/app/(home)/portfolio/playlist/[playlist_slug]/page.tsx
@@ -1,9 +1,7 @@
-import { ProjectCard } from '@/components/cards/projectCard';
-import { PlaylistHeader } from '@/components/header/playlist/playlistHeader';
-import { PlaylistModel } from '@/models/playlistModel';
-import { Playlist } from '@/sanity/schema/schema-types';
-import { Routes } from '@/types/routes';
-import { redirect } from 'next/navigation';
+import { PlaylistPageProvider } from '@/components/layout/playlistPageProvider';
+import { Skeleton } from '@/components/ui/skeleton';
+import { LoaderCircle } from 'lucide-react';
+import { Suspense } from 'react';
 
 // TODO: add the side bar to all the pages
 
@@ -11,25 +9,10 @@ type PlaylistPageProps = {
   params: Promise<{ playlist_slug: string }>
 }
 
-//todo: move this to a constants file
-const HOME_ROUTE: Routes = '/portfolio';
 
 async function PlaylistPage({ params }: PlaylistPageProps) {
-  let playlist: Playlist = null;
 
-  const { playlist_slug } = await params;
-  const isLikedPlaylist = playlist_slug === 'liked-projects';
-  const playlistModel = new PlaylistModel()
-
-  if (isLikedPlaylist) {
-    playlist = await playlistModel.getLikedPlaylist();
-  } else {
-    playlist = await playlistModel.getPlaylist(playlist_slug);
-  }
-
-  if (!playlist || playlist.playlist.length === 0) {
-    return redirect(HOME_ROUTE);
-  }
+  const { playlist_slug: playlistSlug } = await params;
 
   return (
     <div className="flex-1 overflow-hidden relative">
@@ -40,27 +23,16 @@ async function PlaylistPage({ params }: PlaylistPageProps) {
               Page still under construction 👷🏽
             </a>
           </div>
-          <PlaylistHeader title={playlist.playlist_name} />
-          <div
-            className="grid gap-1 sm:gap-4 justify-start py-2 sm:px-10 [grid-template-columns:repeat(auto-fill,minmax(180px,1fr))] 
-            [grid-auto-rows:min-content]"
-          >
-            {playlist.playlist.map((project, i) => (
-              <div key={i} className="size-full flex items-center justify-center">
-                <ProjectCard
-                  key={i}
-                  cardDetails={{
-                    title: `${project.title}`,
-                    slug: project.slug,
-                    first_image_url: project.first_image_url,
-                    sub_title: `${project.sub_title}`,
-                    project_type: project.project_type,
-                  }}
-                />
-              </div >
-            ))
-            }
-          </div>
+
+          {/* duplicate suspense code */}
+          <Suspense
+            fallback={
+              <Skeleton className="w-full h-full bg-zinc-900 flex items-center justify-center">
+                <LoaderCircle className="animate-spin" />
+              </Skeleton>
+            }>
+            <PlaylistPageProvider playlistSlugs={playlistSlug} />
+          </Suspense>
         </div>
       </div>
     </div>
