+++ src/components/input/searchBar.tsx
@@ -3,7 +3,6 @@ import React, { useEffect, useRef } from 'react'
 import { Input } from '../ui/input'
 import { SearchModal } from '../modal/searchModal'
 import { Command, Search, XIcon } from 'lucide-react'
-// import { useCommandKListener } from '@/actions/client-functions/keyStrokes'
 import { Button } from '../ui/button'
 import { usePathname } from 'next/navigation'
 import { CategorisedProject } from '@/sanity/schema/schema-types'
