+++ src/components/layout/resizableLayout.tsx
@@ -3,10 +3,9 @@ import { getCookie } from "@/actions/cookies/cookieHelper";
 import { SidebarHandle } from "@/components/sidebar/sideBarHandle";
 import { SideBar } from "@/components/sidebar/sidebar";
 import { ResizablePanel, ResizablePanelGroup } from "@/components/ui/resizable";
-import { PlaylistModel } from "@/models/playlistModel";
-import { SidebarProvider } from "@/redux/provider/sidebarProvider";
 import clsx from "clsx";
-import React from 'react'
+import React, { Suspense } from 'react'
+import { Skeleton } from "../ui/skeleton";
 
 type ResizableLayoutProps = {
   className: string;
@@ -14,11 +13,7 @@ type ResizableLayoutProps = {
 }
 async function ResizableLayout({ className, children }: ResizableLayoutProps) {
   const DEFAULT_LAYOUT = [20, 80];
-
-  //todo: use promise.all
   const layoutPanes = await getCookie<number[] | null>('react-resizable-panels:layout') || DEFAULT_LAYOUT;
-  const playlists = await new PlaylistModel().getPlaylistsSummary() || [];
-
   return (
     <div className={clsx(className)}>
       <ResizablePanelGroup direction="horizontal">
@@ -26,12 +21,14 @@ async function ResizableLayout({ className, children }: ResizableLayoutProps) {
         <SidebarHandle />
 
 
-        <SidebarProvider playlists={playlists}>
-
-          <SideBar defaultLayout={layoutPanes} className="flex-col h-full bg-zinc-900 rounded-2xl ml-2 mr-1" />
-
-        </SidebarProvider>
-
+        {/* duplicate tailwind css  */}
+        <Suspense fallback={
+          <Skeleton className="flex-col h-full bg-zinc-900 rounded-2xl ml-2 mr-1"
+            style={{ width: `${layoutPanes[0]}%` }}>
+          </Skeleton>
+        }>
+          <SideBar defaultLayout={layoutPanes} />
+        </Suspense>
 
         {/* Main content - Resizable */}
         <ResizablePanel
