+++ src/lib/dev/projectsGenerator.ts
@@ -1,8 +1,9 @@
 // this is used as a mock data generator for projects and blogs for development purposes
 
 import { CategorisedProjects, Project, ProjectTypes } from "@/sanity/schema/schema-types";
+import { sleep } from "../utils";
 
-
+const DELAY_CALL_DEV_TEST = 0
 const images = ['/bart-simpson-cartoon.png', '/mona-lisa.png'];
 
 const randomSlug = (base: string, i: number) => `${base}-${i + 1}`;
@@ -17,13 +18,21 @@ const createRandomProjects = (type: ProjectTypes, count: number) =>
     project_type: type,
   }));
 
-const randomCategorisedProjects: CategorisedProjects = {
-  projects: createRandomProjects('projects', 3),
-  blogs: createRandomProjects('blogs', 6),
-  work_experience: createRandomProjects('work_experience', 2)
+const randomCategorisedProjects = async (): Promise<CategorisedProjects> => {
+  if (process.env.THROTTLE_REQUEST) {
+    await sleep(DELAY_CALL_DEV_TEST)
+  }
+  return {
+    projects: createRandomProjects('projects', 3),
+    blogs: createRandomProjects('blogs', 6),
+    work_experience: createRandomProjects('work_experience', 2)
+  }
 };
 
-const randomProject = (slug: string): Project => {
+const randomProject = async (slug: string): Promise<Project> => {
+  if (process.env.THROTTLE_REQUEST) {
+    await sleep(DELAY_CALL_DEV_TEST)
+  }
   return {
     title: `Sample Project ${slug}`,
     project_images: ["/bart-simpson-cartoon.png", "/mona-lisa.png"],
