+++ src/redux/provider/footerProvider.tsx
@@ -15,6 +15,7 @@ function FooterProvider({ children, projectsArray, currentProject, shuffleEnable
   StoreSingleton.getInstance().dispatch(setProjectsArray(projectsArray))
   StoreSingleton.getInstance().dispatch(setShuffle(shuffleEnabled))
 
+
   return (
     <>{children}</>
   )
